ofers create and delete

+++ frontend/src/app/client-schedule/client-schedule.component.ts
@@ -4,6 +4,7 @@ import { ActivatedRoute } from '@angular/router';
 import { Client } from 'src/model/client.modul';
 import { AnnotationsService } from '../annotations.service';
 import { ClientsService } from '../clients.service';
+import { OffersService } from '../offers.service';
 
 @Component({
   selector: 'app-client-schedule',
@@ -12,26 +13,25 @@ import { ClientsService } from '../clients.service';
 })
 export class ClientScheduleComponent implements OnInit {
   clientID: any;
-  client: any;
-  annotations: any;
-  modifyClientForm!: FormGroup;
+  client!: Client;
+  annotations: Array<any> = [];
+  offers: Array<any> = [];
 
   constructor(
     private router: ActivatedRoute,
     private clientsService: ClientsService,
-    private annotationsSevice: AnnotationsService
+    private annotationsSevice: AnnotationsService,
+    private offersService: OffersService
   ) {}
 
-  private initForm() {
-    this.modifyClientForm = new FormGroup({
-      name: new FormControl(),
-      surname: new FormControl(),
-      company_name: new FormControl(),
-      email: new FormControl(),
-      p_iva: new FormControl(),
-      phone: new FormControl(),
-    });
-  }
+  modifyClientForm = new FormGroup({
+    name: new FormControl(),
+    surname: new FormControl(),
+    business_name: new FormControl(),
+    email: new FormControl(),
+    p_iva: new FormControl(),
+    phone: new FormControl(),
+  });
 
   // get all the data
   getData() {
@@ -40,12 +40,28 @@ export class ClientScheduleComponent implements OnInit {
       .subscribe((res: any) => {
         this.client = res.client;
         this.annotations = res.annotations;
+        this.offers = res.offers;
+        this.modifyClientForm.patchValue(this.client);
       });
   }
 
-  deleteAnnotation(id: number) {
-    this.annotationsSevice.deleteAnnotation(id);
+  modifyClient() {
+    const data = this.modifyClientForm.value;
+    this.clientsService.modifyClient(data, this.client.id);
     this.getData();
+  }
+
+  deleteAnnotation(id: number) {
+    if (confirm("Sei sicuro di voler elimanre l'annotazione ?")) {
+      this.annotationsSevice.deleteAnnotation(id);
+      this.getData();
+    }
+  }
+  deleteOffer(id: number) {
+    if (confirm('Sei sicuro di voler elimanre questa offerta?')) {
+      this.offersService.deleteOffer(id);
+      this.getData();
+    }
   }
 
   update(event: string) {
@@ -55,6 +71,5 @@ export class ClientScheduleComponent implements OnInit {
   ngOnInit(): void {
     this.clientID = this.router.snapshot.paramMap.get('id');
     this.getData();
-    this.initForm();
   }
 }
